pyproject.toml
[project]
name = "airflow-demo"
version = "0.1.0"
description = ""
authors = [
    {name = "Xavier Mandeng",email = "mandeng@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.12"
dependencies = [
    "apache-airflow (==2.8.3)",
    "cryptography (>=44.0.2,<45.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
